+++ NEWS.md
@@ -1,3 +1,30 @@
+# celltrackR 1.2.0
+
+Integration with immunemap.org. 
+
+### Potentially breaking changes: 
+
+None.
+
+### Minor changes:
+New functionalities for integration with immunemap.org:
+- (immunemap-integration.R => read.immap.json ): new function, reads data from an 
+	immunemap json file (or an url pointing to such a file). This returns a list with
+	both track object(s) and metadata; see documentation ?read.immap.json. 
+- (immunemap-integration.R) : see above; file contains also several helper functions.
+- tests added for these reading functions as well.
+- (inst/extdata/immunemap.json) : new example file showing the format for immunemap json.
+
+
+### Documentation changes:
+
+Minor changes:
+- including documentation for new functionalities for integration with immunemap.org;
+- Updated package documentation after breaking change in Roxygen upon CRAN request.
+- Updated the documentation for the included datasets, which had not yet been updated 
+with the new versions of the data included since package v 1.0.0.
+
+
 # celltrackR 1.1.0
 
 Significant speed-up in some of the angle analysis functions.